+++ src/components/forms/RangeInput.tsx
@@ -1,0 +1,95 @@
+import {
+  ChildFunctionProps,
+} from "kea-forms";
+import React, { useContext } from "react";
+import styled from "styled-components";
+import {
+  Box,
+  Button,
+  RangeInput as GRangeInput,
+  Paragraph,
+  ResponsiveContext,
+} from "grommet";
+import { Subtract, Add } from "grommet-icons";
+import { InputProps } from "./types";
+
+interface NumberInputProps extends InputProps {
+  min: number | `${number}`;
+  max: number | `${number}`;
+  step: number | `${number}`;
+}
+
+const RangeInputContainer = styled(Box)`
+  container-type: inline-size;
+  container-name: range-input-container;
+
+  .button {
+    display: none;
+    border-radius: 100%;
+  }
+
+  @container range-input-container (width > 300px) {
+    .button {
+      display: revert;
+    }
+  }
+`;
+
+export function RangeInput({
+  value,
+  onChange,
+  min,
+  max,
+  step,
+}: Partial<ChildFunctionProps> & NumberInputProps) {
+  const size = useContext(ResponsiveContext);
+
+  return (
+    <RangeInputContainer
+      direction="row"
+      align="start"
+      pad="small"
+      gap={"small"}
+    >
+      <Button
+        className="button"
+        disabled={Number(value) <= Number(min)}
+        icon={<Subtract color="control" />}
+        onClick={() => {
+          onChange && onChange(Number(value) - Number(step));
+        }}
+        size={size === "small" ? "small" : "medium"}
+        alignSelf="start"
+        secondary
+        hoverIndicator={{ elevation: "small", background: {} }}
+      />
+      <Box align="center" width="small" flex direction="column">
+        <GRangeInput
+          a11yTitle="Select range value"
+          min={min}
+          max={max}
+          step={Number(step)}
+          color="control"
+          value={String(value)}
+          onChange={({ target: { value } }) => onChange && onChange(value)}
+        />
+        <Box direction="row" fill justify="between">
+          <Paragraph size="small">{min}</Paragraph>
+          <Paragraph size="small">{max}</Paragraph>
+        </Box>
+      </Box>
+      <Button
+        className="button"
+        disabled={Number(value) >= Number(max)}
+        icon={<Add color="control" />}
+        onClick={() => {
+          onChange && onChange(Number(value) + Number(step));
+        }}
+        size={size === "small" ? "small" : "medium"}
+        alignSelf="start"
+        secondary
+        hoverIndicator={{ elevation: "small", background: {} }}
+      />
+    </RangeInputContainer>
+  );
+}
