+++ src/components/forms/types.d.ts
@@ -1,0 +1,4 @@
+export interface InputProps {
+  name: string;
+  id: string;
+}
